Simplify analytics settings component.
