+++ assets/js/modules/analytics/components/settings/UASettingsView.js
@@ -20,7 +20,7 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { createInterpolateElement } from '@wordpress/element';
+import { createInterpolateElement, Fragment } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -51,48 +51,103 @@ export default function UASettingsView() {
 			path: escapeURI`/a${ accountID }w${ internalWebPropertyID }p${ profileID }/admin/view/settings`,
 		} )
 	);
+	const useSnippet = useSelect( ( select ) =>
+		select( MODULES_ANALYTICS ).getUseSnippet()
+	);
+	const canUseSnippet = useSelect( ( select ) =>
+		select( MODULES_ANALYTICS ).getCanUseSnippet()
+	);
+	const hasExistingTag = useSelect( ( select ) =>
+		select( MODULES_ANALYTICS ).hasExistingTag()
+	);
 
 	return (
-		<div className="googlesitekit-settings-module__meta-items">
-			<div className="googlesitekit-settings-module__meta-item">
-				<h5 className="googlesitekit-settings-module__meta-item-type">
-					{ __( 'Account', 'google-site-kit' ) }
-				</h5>
-				<p className="googlesitekit-settings-module__meta-item-data">
-					<DisplaySetting value={ accountID } />
-				</p>
+		<Fragment>
+			<div className="googlesitekit-settings-module__meta-items">
+				<div className="googlesitekit-settings-module__meta-item">
+					<h5 className="googlesitekit-settings-module__meta-item-type">
+						{ __( 'Account', 'google-site-kit' ) }
+					</h5>
+					<p className="googlesitekit-settings-module__meta-item-data">
+						<DisplaySetting value={ accountID } />
+					</p>
+				</div>
+				<div className="googlesitekit-settings-module__meta-item">
+					<h5 className="googlesitekit-settings-module__meta-item-type">
+						{ __( 'Property', 'google-site-kit' ) }
+					</h5>
+					<p className="googlesitekit-settings-module__meta-item-data">
+						<DisplaySetting value={ propertyID } />
+					</p>
+				</div>
+				<div className="googlesitekit-settings-module__meta-item">
+					<h5 className="googlesitekit-settings-module__meta-item-type">
+						{ __( 'View', 'google-site-kit' ) }
+					</h5>
+					<p className="googlesitekit-settings-module__meta-item-data">
+						<DisplaySetting value={ profileID } />
+					</p>
+				</div>
+				<div className="googlesitekit-settings-module__meta-item googlesitekit-settings-module__meta-item--data-only">
+					<p className="googlesitekit-settings-module__meta-item-data googlesitekit-settings-module__meta-item-data--tiny">
+						<Link href={ editViewSettingsURL } external>
+							{ createInterpolateElement(
+								__(
+									'Edit <VisuallyHidden>Universal Analytics property view </VisuallyHidden>in Analytics',
+									'google-site-kit'
+								),
+								{
+									VisuallyHidden: <VisuallyHidden />,
+								}
+							) }
+						</Link>
+					</p>
+				</div>
 			</div>
-			<div className="googlesitekit-settings-module__meta-item">
-				<h5 className="googlesitekit-settings-module__meta-item-type">
-					{ __( 'Property', 'google-site-kit' ) }
-				</h5>
-				<p className="googlesitekit-settings-module__meta-item-data">
-					<DisplaySetting value={ propertyID } />
-				</p>
-			</div>
-			<div className="googlesitekit-settings-module__meta-item">
-				<h5 className="googlesitekit-settings-module__meta-item-type">
-					{ __( 'View', 'google-site-kit' ) }
-				</h5>
-				<p className="googlesitekit-settings-module__meta-item-data">
-					<DisplaySetting value={ profileID } />
-				</p>
-			</div>
-			<div className="googlesitekit-settings-module__meta-item googlesitekit-settings-module__meta-item--data-only">
-				<p className="googlesitekit-settings-module__meta-item-data googlesitekit-settings-module__meta-item-data--tiny">
-					<Link href={ editViewSettingsURL } external>
-						{ createInterpolateElement(
-							__(
-								'Edit <VisuallyHidden>Universal Analytics property view </VisuallyHidden>in Analytics',
-								'google-site-kit'
-							),
-							{
-								VisuallyHidden: <VisuallyHidden />,
-							}
+			<div className="googlesitekit-settings-module__meta-items">
+				<div className="googlesitekit-settings-module__meta-item">
+					<h5 className="googlesitekit-settings-module__meta-item-type">
+						{ __(
+							'Universal Analytics Code Snippet',
+							'google-site-kit'
 						) }
-					</Link>
-				</p>
+					</h5>
+					<p className="googlesitekit-settings-module__meta-item-data">
+						{ canUseSnippet === false && (
+							<span>
+								{ __(
+									'The code is controlled by the Tag Manager module',
+									'google-site-kit'
+								) }
+							</span>
+						) }
+						{ canUseSnippet && useSnippet && (
+							<span>
+								{ __(
+									'Snippet is inserted',
+									'google-site-kit'
+								) }
+							</span>
+						) }
+						{ canUseSnippet && ! useSnippet && ! hasExistingTag && (
+							<span>
+								{ __(
+									'Snippet is not inserted',
+									'google-site-kit'
+								) }
+							</span>
+						) }
+						{ canUseSnippet && ! useSnippet && hasExistingTag && (
+							<span>
+								{ __(
+									'Inserted by another plugin or theme',
+									'google-site-kit'
+								) }
+							</span>
+						) }
+					</p>
+				</div>
 			</div>
-		</div>
+		</Fragment>
 	);
 }
