+++ assets/js/modules/analytics/components/settings/index.js
@@ -23,9 +23,6 @@ export { default as SettingsControls } from './SettingsControls';
 export { default as SettingsUseSnippetSwitch } from './SettingsUseSnippetSwitch';
 export { default as GA4SettingsControls } from './GA4SettingsControls';
 export { default as GA4SettingsNotice } from './GA4SettingsNotice';
-export { default as AddConversionIDSettingsView } from './AddConversionIDSettingsView';
-export { default as ExcludeFromAnalyticsSettingsView } from './ExcludeFromAnalyticsSettingsView';
+export { default as OptionalSettingsView } from './OptionalSettingsView';
 export { default as GA4SettingsView } from './GA4SettingsView';
-export { default as IPAnonymizationSettingsView } from './IPAnonymizationSettingsView';
 export { default as UASettingsView } from './UASettingsView';
-export { default as UASnippetSettingsView } from './UASnippetSettingsView';
